Refactor confidence from replay stage
