+++ core/src/confidence.rs
@@ -1,4 +1,10 @@
+use crate::consensus::StakeLockout;
+use crate::service::Service;
 use std::collections::{HashMap, HashSet};
+use std::sync::atomic::{AtomicBool, Ordering};
+use std::sync::mpsc::{channel, Receiver, Sender};
+use std::sync::{Arc, RwLock};
+use std::thread::{self, Builder, JoinHandle};
 
 #[derive(Debug, Default, PartialEq)]
 pub struct Confidence {
@@ -82,6 +88,131 @@ impl ForkConfidenceCache {
     }
 }
 
+pub struct ConfidenceAggregationData {
+    lockouts: HashMap<u64, StakeLockout>,
+    root: Option<u64>,
+    ancestors: Arc<HashMap<u64, HashSet<u64>>>,
+    total_staked: u64,
+}
+
+impl ConfidenceAggregationData {
+    pub fn new(
+        lockouts: HashMap<u64, StakeLockout>,
+        root: Option<u64>,
+        ancestors: Arc<HashMap<u64, HashSet<u64>>>,
+        total_staked: u64,
+    ) -> Self {
+        Self {
+            lockouts,
+            root,
+            ancestors,
+            total_staked,
+        }
+    }
+}
+
+pub struct AggregateConfidenceService {
+    t_confidence: JoinHandle<()>,
+}
+
+impl AggregateConfidenceService {
+    pub fn aggregate_confidence(
+        root: Option<u64>,
+        ancestors: &HashMap<u64, HashSet<u64>>,
+        stake_lockouts: &HashMap<u64, StakeLockout>,
+    ) -> HashMap<u64, u128> {
+        let mut stake_weighted_lockouts: HashMap<u64, u128> = HashMap::new();
+        for (fork, lockout) in stake_lockouts.iter() {
+            if root.is_none() || *fork >= root.unwrap() {
+                let mut slot_with_ancestors = vec![*fork];
+                slot_with_ancestors.extend(ancestors.get(&fork).unwrap_or(&HashSet::new()));
+                for slot in slot_with_ancestors {
+                    if root.is_none() || slot >= root.unwrap() {
+                        let entry = stake_weighted_lockouts.entry(slot).or_default();
+                        *entry += u128::from(lockout.lockout()) * u128::from(lockout.stake());
+                    }
+                }
+            }
+        }
+        stake_weighted_lockouts
+    }
+
+    pub fn new(
+        exit: &Arc<AtomicBool>,
+        fork_confidence_cache: Arc<RwLock<ForkConfidenceCache>>,
+    ) -> (Sender<ConfidenceAggregationData>, Self) {
+        let (lockouts_sender, lockouts_receiver): (
+            Sender<ConfidenceAggregationData>,
+            Receiver<ConfidenceAggregationData>,
+        ) = channel();
+        let exit_ = exit.clone();
+        (
+            lockouts_sender,
+            Self {
+                t_confidence: Builder::new()
+                    .name("solana-aggregate-stake-lockouts".to_string())
+                    .spawn(move || loop {
+                        if exit_.load(Ordering::Relaxed) {
+                            break;
+                        }
+                        if let Ok(aggregation_data) = lockouts_receiver.try_recv() {
+                            let stake_weighted_lockouts = Self::aggregate_confidence(
+                                aggregation_data.root,
+                                &aggregation_data.ancestors,
+                                &aggregation_data.lockouts,
+                            );
+
+                            let mut w_fork_confidence_cache =
+                                fork_confidence_cache.write().unwrap();
+
+                            // Cache the confidence values
+                            for (fork, stake_lockout) in aggregation_data.lockouts.iter() {
+                                if aggregation_data.root.is_none()
+                                    || *fork >= aggregation_data.root.unwrap()
+                                {
+                                    w_fork_confidence_cache.cache_fork_confidence(
+                                        *fork,
+                                        stake_lockout.stake(),
+                                        aggregation_data.total_staked,
+                                        stake_lockout.lockout(),
+                                    );
+                                }
+                            }
+
+                            // Cache the stake weighted lockouts
+                            for (fork, stake_weighted_lockout) in stake_weighted_lockouts.iter() {
+                                if aggregation_data.root.is_none()
+                                    || *fork >= aggregation_data.root.unwrap()
+                                {
+                                    w_fork_confidence_cache.cache_stake_weighted_lockouts(
+                                        *fork,
+                                        *stake_weighted_lockout,
+                                    )
+                                }
+                            }
+
+                            if let Some(root) = aggregation_data.root {
+                                w_fork_confidence_cache
+                                    .prune_confidence_cache(&aggregation_data.ancestors, root);
+                            }
+
+                            drop(w_fork_confidence_cache);
+                        }
+                    })
+                    .unwrap(),
+            },
+        )
+    }
+}
+
+impl Service for AggregateConfidenceService {
+    type JoinReturnType = ();
+
+    fn join(self) -> thread::Result<()> {
+        self.t_confidence.join()
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -118,5 +249,31 @@ mod tests {
             cache.confidence.get(&fork).unwrap().stake_weighted_lockouts,
             20,
         );
+    }
+
+    #[test]
+    fn test_aggregate_confidence() {
+        let stakes = vec![
+            (0, StakeLockout::new(1, 32)),
+            (1, StakeLockout::new(1, 24)),
+            (2, StakeLockout::new(1, 16)),
+            (3, StakeLockout::new(1, 8)),
+        ]
+        .into_iter()
+        .collect();
+        let ancestors = vec![
+            (0, HashSet::new()),
+            (1, vec![0].into_iter().collect()),
+            (2, vec![0, 1].into_iter().collect()),
+            (3, vec![0, 1, 2].into_iter().collect()),
+        ]
+        .into_iter()
+        .collect();
+        let stake_weighted_lockouts =
+            AggregateConfidenceService::aggregate_confidence(Some(1), &ancestors, &stakes);
+        assert!(stake_weighted_lockouts.get(&0).is_none());
+        assert_eq!(*stake_weighted_lockouts.get(&1).unwrap(), 8 + 16 + 24);
+        assert_eq!(*stake_weighted_lockouts.get(&2).unwrap(), 8 + 16);
+        assert_eq!(*stake_weighted_lockouts.get(&3).unwrap(), 8);
     }
 }
